requirements.txt
wheel
attrs==23.1.0
blessed==1.20.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorful==0.5.5
contourpy==1.2.0
frozenlist==1.4.0
grpcio==1.60.0
huggingface-hub==0.19.4
idna==3.6
imageio==2.33.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
munch==4.0.0
networkx==3.2.1
numpy==1.26.2
opencv-python-headless==4.8.1.78
pandas==2.1.4
Pillow==10.1.0
pix2tex==0.1.2
py-spy==0.3.14
pycwt==0.4.0b0
pyparsing==3.1.1
pyperclip==1.8.2
pytz==2023.3.post1
PyWavelets==1.5.0
PyYAML==6.0.1
ray[all]
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
rsa==4.9
SciencePlots==2.1.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
shiboken6==6.6.1
six==1.16.0
sympy==1.12
tifffile==2023.12.9
timm==0.5.4
tokenizers==0.15.0
triton==2.1.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
x-transformers==0.15.0
xlrd==2.0.1
yarl==1.9.4
zipp==3.17.0
notebook
jupyterlab
torch
torchaudio
torchvision
tqdm==4.66.1
transformers==4.35.2
flask
ipywidgets
ipykernel
jupyterlab-git
jupyterlab-lsp
jupyterlab-toc
